Update patch set 4
Patch Set 4:

I get the same WallpaperPickerActivity exception that David gets and also one for the LockWallpaperPickerActivity: 
http://pastebin.com/raw.php?i=fbzXD8Wd

The icon settings print this http://pastebin.com/raw.php?i=jgpYEjqU in the logcat. Seems to be a bit overkill to show the complete trace. And because cameranext is not installed the camera icon is replaced by the Android robot.

I have no themes installed and it's showing "Holo (Default)" everywhere.

Patch-set: 4