Update patch set 5
Patch Set 5:

@Thomas & Danny - Your wallpaper issue is unrelated to this patch. Your trebuchet (or a subset of Trebuchet) is outdated, disabled, or removed. A regular CM build will work fine with this patch.

In any case this issue unrelated to the chooser patch. Wallpaper picker is launched from Settings, so if you want to fix it by making a more general intent (in Settings) or separating WallpaperPickerActivity from Trebuchet we can support you by creating a new repo and reviewing it, but lets not hold up this patch because of it.

Patch-set: 5