Update patch set 9
Patch Set 9:

It's an easy thing: Can Trebuchet be deactivated via Settings -> Apps? If yes, we shouldn't build core functionality into it. If no, it's no big deal.

(This has nothing to do with the patch in question, BTW)

(Yes, I also think making the lockscreen wallpaper picker depend on Trebuchet being present is not a good idea.)

Patch-set: 9