Update patch set 9
Patch Set 9:

@Kevin for that i have uploaded the following commit http://review.cyanogenmod.org/#/c/63576/

it catches if trebuchet is not installed and also do not explicitly tries to start the activity via packagename and classname, rather it uses the intent action.

if the wallpaperpicker will be moved around or whatever in future, it will still be working correctly, if the new activity can be launched via the same action.

Patch-set: 9
Reviewer: Alexander Martinz <8328@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0