Update patch set 9
Patch Set 9:

@Danny That's just the thing Trebuchet cant be disabled in App info.. 

@Kevin No its not a pissing match and its not about if I agree or disagree that the wallpaper chooser be tied to Trebuchet. 

It's the fact that, as someone who has developed ROM's for years now, I do stand strong on the fact that applications in the system/apps folder shouldn't just be plucked out. If they are and something breaks the person who did it really shouldn't then go complaining to the developer.

@Alexander That looks like a resolution to the current complaint and is how thing probably should have been done upfront thanks for the commit. 

While I agree that the fix Alexander committed is a better solution it doesn't change my stance that people nor gapps should be randomly removing application from the /syste/apps folder

Patch-set: 9