Update patch set 9
Patch Set 9:

Wow this should be an easy one guys, sorry to intrude but there are so many users that the very first thing they do is remove the stock launcher for space. Other roms might not want to include it as well. If it's not too much trouble to move it to framework for everyone to be happy then why is this an issue? If there's something that's critical keeping it in Trebuchet then there should be a way to warn users on first boot not too remove. Same with Rom cooks and developers. 

Sorry just wanted to try and help. 

DJ

Patch-set: 9
Reviewer: Bryan Owens <5959@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0