Update patch set 9
Patch Set 9:

Oh you'd be surprised what users do that makes no sense, flashing themed apps to system uses space as well. But the fact is people do it just to remove it as an eye sore or even just because in the past they could without issue. All I'm saying is if you make this change without needing to then why do it? If you have to the just warn them. Been a user/Themer of cm since the G1 so I've seen this a ton.

Patch-set: 9