Update patch set 9
Patch Set 9:

I personally don't use it but keep it on the system. But users...You gotta consider them "for the lack of a better term" stupid to the fact that...If this can be gone remove it....They think root is just for that...Less on system faster it goes. CM has always done things for the greater good. Unless it's a critical thing then why do it? You can see it picking a launcher. That's a user seeing the icon and not wanting to..... Anymore.

Patch-set: 9