Update patch set 9
Patch Set 9: -Code-Review

This argument is over already but I do want to add just my 0.02c

We used to take great care (and pride) to ensure that core functionality of CM was not end-user breakable (look at how we implemented Blacklist) and, while we had some derps from time to time, very rarely did we end up with stuff that was really broken.  Defensive coding is a good thing and one should always assume that users are going to do the worst and build the appropriate fail-safes.

Launchers are a very personal choice for users and you cannot make a core feature of one part of the system depend on something as 'transient'.

I have not been personally very active in CM dev for the past few months and there certainly seems to be a bit more of a 'cowboy' attitude towards commits in CM lately...

Patch-set: 9
Label: Code-Review=0