Update patch set 9
Patch Set 9:

@Christopher thats because of the new policy.
you need to override isValidFragment and return the classname of the fragment if you have a preference activity with preference headers.

this mostly applies for tablets as they use it in multi pane mode.

this valid fragment restriction is because of security as before any malicious app could inject fragments and hijack or whatever.

Patch-set: 9