Update patch set 1
Patch Set 1:

It does what it is supposed to do. But I'd prefer a additional toast message telling you that the theme was uninstalled. 
An alternative to closing the ThemeChooser could be to go back to the theme overview (also with a toast message).

After tapping on the notification you are able to repeatedly open the ThemeChooser with the uninstalled package name through the recent apps. Which will close it immediately again.

Patch-set: 1