Update patch set 1
Patch Set 1: Code-Review-1

I passed this on to our UX guys and they like what this brings.  They still want the user to be able to uncheck an item and have that component go back to default when applied.  We are going to work on getting that functionality working in the chooser and frameworks.  Once that is done we can revisit this change, which will most likely need to be rebased and a few things changed.

Thanks for this contribution and pleas hang tight.

Patch-set: 1