Update patch set 8
Patch Set 8:

Feedback 3:

As you can se in the same screenshot (https://www.dropbox.com/s/w4rvi6r98c9qe10/Screenshot_2014-05-16-13-50-32.png), wallpapers images are not shown most of the time. You ony get a grey background.

The wallpaper _is_ shown on the main page of Theme Chooser, but not in the detailed view of that particular theme.

More themes that have this problem, to name a few:

- https://play.google.com/store/apps/details?id=com.botty.theme.next


- https://play.google.com/store/apps/details?id=com.hoppermi.themenext.darkergreen


- https://play.google.com/store/apps/details?id=org.cyanogenmod.great.freedom

Patch-set: 8