Update patch set 8
Patch Set 8:

Feedback 8:

When selecting Theme packs and using the page indicators, I am confused as to what indicator is what.

There are 9 options to choose from in a theme pack, but only 6 page indicators. What corresponds to what?

1) Style - page 1
2) Wallpapers - page 2
3) Lock screen - NO page?
4) Fonts - page 3
5) Icons - NO page? I do want to see a icon preview!
6) Boot animations - page 6
7) Ringtones - page 4 (huh?)
8) Notifications - page 4 again
9) Alarms - page 4 again


Soooo...what does page 5 do? It's just showing the wallpaper again? 







That conludes my feedback for now and I would appreciate a response if it was helpfull.

Patch-set: 8