Update patch set 6
Patch Set 6: Code-Review-1

After speaking with some UX guys, it would be better to change the available label to indicate to the user that they have applied all available items from this theme.  In the meantime we should go back to the previous patch set and then create a new patch for dealing with this particular issue.

Patch-set: 6