Update patch set 3
Patch Set 3:

Themes Showcase was unpublished from the Play Store.  That means the only way a user is going to find it is by using a direct link to it.  Try https://play.google.com/store/search?q=themes%20showcase and see if you can find it ;)

The direct link was removed because the app was unpublished, and it was unpublished because it was no longer being updated with CM12 and CM12.1 themes.  With that said, this patch is valid.

Patch-set: 3