Update patch set 6
Patch Set 6:

Seeing as the capps case was his point, that package will replace both dialer and chooser, so this being in the OSS code is unnecessary. 

Adding DialerNext here would only be necessary if the user flashed a custom Capps, which we shouldn't spiral down the slope of supporting.

Patch-set: 6