Update patch set 9
Patch Set 9:

Something I did not convey with my last comment is I see App themer as it's own brand and therefore should have iconography that is unique to it but still has a connection with the theme chooser's brand as well.

With the palette icon being the main branding for the chooser, maybe app themer needs an artists brush instead of a wide paint brush used for painting walls and fences. ;)

Patch-set: 9